requirements.txt
advertools
emoji
matplotlib
numpy
pandas
Pillow
seaborn
streamlit
urlextract
wordcloud
